+++ src/Gaufrette/Filesystem/Adapter/Sftp.php
@@ -41,8 +41,11 @@ class Sftp implements Adapter
      */
     public function rename($key, $new)
     {
-        $this->write($new, $this->read($key));
-        $this->delete($key);
+        if (!$this->write($new, $this->read($key))) {
+            $this->delete($key);
+        }
+
+        return false;
     }
 
     /**
